+++ openapi_client/models/timeseries_rain.py
@@ -162,7 +162,7 @@ class TimeseriesRain(object):
     def duration(self):
         """Gets the duration of this TimeseriesRain.  # noqa: E501
 
-        Duration of event in seconds  # noqa: E501
+        event duration in seconds. -9999 is the 'infinite duration' value (only allowed in conjunction with infinite simulations  # noqa: E501
 
         :return: The duration of this TimeseriesRain.  # noqa: E501
         :rtype: int
@@ -173,7 +173,7 @@ class TimeseriesRain(object):
     def duration(self, duration):
         """Sets the duration of this TimeseriesRain.
 
-        Duration of event in seconds  # noqa: E501
+        event duration in seconds. -9999 is the 'infinite duration' value (only allowed in conjunction with infinite simulations  # noqa: E501
 
         :param duration: The duration of this TimeseriesRain.  # noqa: E501
         :type: int